Move sourcecode into subdirectory /src
