+++ OP2Utility.vcxproj
@@ -115,58 +115,58 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="Archives\ArchivePacker.h" />
-    <ClInclude Include="Archives\ArchiveUnpacker.h" />
-    <ClInclude Include="Archives\CompressionType.h" />
-    <ClInclude Include="Archives\WaveFile.h" />
+    <ClInclude Include="src\Archives\ArchivePacker.h" />
+    <ClInclude Include="src\Archives\ArchiveUnpacker.h" />
+    <ClInclude Include="src\Archives\CompressionType.h" />
+    <ClInclude Include="src\Archives\WaveFile.h" />
     <ClInclude Include="include\OP2Utility.h" />
-    <ClInclude Include="Maps\CellType.h" />
-    <ClInclude Include="Maps\ClipRect.h" />
-    <ClInclude Include="Maps\MapData.h" />
-    <ClInclude Include="Maps\MapHeader.h" />
-    <ClInclude Include="Maps\MapReader.h" />
-    <ClInclude Include="Maps\MapWriter.h" />
-    <ClInclude Include="Maps\TerrainType.h" />
-    <ClInclude Include="Maps\TileData.h" />
-    <ClInclude Include="Maps\TileGroup.h" />
-    <ClInclude Include="Maps\TileInfo.h" />
-    <ClInclude Include="Maps\TileSetSource.h" />
-    <ClInclude Include="ResourceManager.h" />
-    <ClInclude Include="Archives\AdaptHuffTree.h" />
-    <ClInclude Include="Archives\ArchiveFile.h" />
-    <ClInclude Include="Archives\BitStream.h" />
-    <ClInclude Include="Archives\ClmFile.h" />
-    <ClInclude Include="Archives\HuffLZ.h" />
-    <ClInclude Include="Archives\VolFile.h" />
-    <ClCompile Include="Archives\ArchivePacker.cpp" />
-    <ClCompile Include="Archives\ArchiveUnpacker.cpp" />
-    <ClCompile Include="Maps\MapData.cpp" />
-    <ClCompile Include="Maps\MapReader.cpp" />
-    <ClCompile Include="Maps\MapWriter.cpp" />
-    <ClCompile Include="ResourceManager.cpp" />
-    <ClCompile Include="Archives\AdaptHuffTree.cpp" />
-    <ClCompile Include="Archives\BitStream.cpp" />
-    <ClCompile Include="Archives\ClmFile.cpp" />
-    <ClCompile Include="Archives\HuffLZ.cpp" />
-    <ClCompile Include="Archives\VolFile.cpp" />
-    <ClCompile Include="Streams\FileSliceReader.cpp" />
-    <ClCompile Include="Streams\FileStreamReader.cpp" />
-    <ClCompile Include="Streams\FileStreamWriter.cpp" />
-    <ClCompile Include="Streams\MemoryStreamWriter.cpp" />
-    <ClCompile Include="Streams\MemoryStreamReader.cpp" />
-    <ClInclude Include="Streams\FileSliceReader.h" />
-    <ClInclude Include="Streams\FileStreamReader.h" />
-    <ClInclude Include="Streams\FileStreamWriter.h" />
-    <ClInclude Include="Streams\MemoryStreamReader.h" />
-    <ClInclude Include="Streams\MemoryStreamWriter.h" />
-    <ClInclude Include="Streams\SeekableStreamReader.h" />
-    <ClInclude Include="Streams\SeekableStreamWriter.h" />
-    <ClInclude Include="Streams\StreamReader.h" />
-    <ClInclude Include="Streams\StreamWriter.h" />
-    <ClInclude Include="StringHelper.h" />
-    <ClInclude Include="XFile.h" />
-    <ClCompile Include="StringHelper.cpp" />
-    <ClCompile Include="XFile.cpp" />
+    <ClInclude Include="src\Maps\CellType.h" />
+    <ClInclude Include="src\Maps\ClipRect.h" />
+    <ClInclude Include="src\Maps\MapData.h" />
+    <ClInclude Include="src\Maps\MapHeader.h" />
+    <ClInclude Include="src\Maps\MapReader.h" />
+    <ClInclude Include="src\Maps\MapWriter.h" />
+    <ClInclude Include="src\Maps\TerrainType.h" />
+    <ClInclude Include="src\Maps\TileData.h" />
+    <ClInclude Include="src\Maps\TileGroup.h" />
+    <ClInclude Include="src\Maps\TileInfo.h" />
+    <ClInclude Include="src\Maps\TileSetSource.h" />
+    <ClInclude Include="src\ResourceManager.h" />
+    <ClInclude Include="src\Archives\AdaptHuffTree.h" />
+    <ClInclude Include="src\Archives\ArchiveFile.h" />
+    <ClInclude Include="src\Archives\BitStream.h" />
+    <ClInclude Include="src\Archives\ClmFile.h" />
+    <ClInclude Include="src\Archives\HuffLZ.h" />
+    <ClInclude Include="src\Archives\VolFile.h" />
+    <ClCompile Include="src\Archives\ArchivePacker.cpp" />
+    <ClCompile Include="src\Archives\ArchiveUnpacker.cpp" />
+    <ClCompile Include="src\Maps\MapData.cpp" />
+    <ClCompile Include="src\Maps\MapReader.cpp" />
+    <ClCompile Include="src\Maps\MapWriter.cpp" />
+    <ClCompile Include="src\ResourceManager.cpp" />
+    <ClCompile Include="src\Archives\AdaptHuffTree.cpp" />
+    <ClCompile Include="src\Archives\BitStream.cpp" />
+    <ClCompile Include="src\Archives\ClmFile.cpp" />
+    <ClCompile Include="src\Archives\HuffLZ.cpp" />
+    <ClCompile Include="src\Archives\VolFile.cpp" />
+    <ClCompile Include="src\Streams\FileSliceReader.cpp" />
+    <ClCompile Include="src\Streams\FileStreamReader.cpp" />
+    <ClCompile Include="src\Streams\FileStreamWriter.cpp" />
+    <ClCompile Include="src\Streams\MemoryStreamWriter.cpp" />
+    <ClCompile Include="src\Streams\MemoryStreamReader.cpp" />
+    <ClInclude Include="src\Streams\FileSliceReader.h" />
+    <ClInclude Include="src\Streams\FileStreamReader.h" />
+    <ClInclude Include="src\Streams\FileStreamWriter.h" />
+    <ClInclude Include="src\Streams\MemoryStreamReader.h" />
+    <ClInclude Include="src\Streams\MemoryStreamWriter.h" />
+    <ClInclude Include="src\Streams\SeekableStreamReader.h" />
+    <ClInclude Include="src\Streams\SeekableStreamWriter.h" />
+    <ClInclude Include="src\Streams\StreamReader.h" />
+    <ClInclude Include="src\Streams\StreamWriter.h" />
+    <ClInclude Include="src\StringHelper.h" />
+    <ClInclude Include="src\XFile.h" />
+    <ClCompile Include="src\StringHelper.cpp" />
+    <ClCompile Include="src\XFile.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
