+++ src/Maps/ClipRect.h
