+++ src/ResourceManager.cpp
