+++ src/ResourceManager.h
