+++ src/Streams/FileSliceReader.h
