+++ src/Streams/FileStreamWriter.cpp
