+++ src/Streams/MemoryStreamReader.cpp
