+++ src/Streams/MemoryStreamReader.h
