+++ src/Streams/MemoryStreamWriter.cpp
