+++ src/Streams/MemoryStreamWriter.h
