+++ src/Streams/SeekableStreamReader.h
