+++ src/Streams/SeekableStreamWriter.h
