+++ src/Streams/StreamReader.h
