+++ src/Streams/StreamWriter.h
