+++ src/StringHelper.cpp
