+++ src/StringHelper.h
